Split groupByFieldsForStatistics by comma
ArcGIS Operations Dashboard allows you to create a Serial Chart,
grouping the data one field, and splitting the bars by a second. The
queries it sends to the feature service are comma separated. This was
being treated by winnow as a single field name containing a comma,
instead of splitting the parameter into a list of field names.

Updated the normalizeGroupBy function to split the parameter with the
comma character. (I've done this for both 'groupBy' and
'groupByFieldsForStatistics', could maybe restrict it to only do it for
the Esri 'groupByFieldsForStatistics' property?)

Added unit test to cover it.